# internal/httpapi/openapi/swagger.yaml
basePath: /api/v1
definitions:
...
  httpapi.CommandRequest:
    properties:
      args:
        description: 'Command and arguments: strings, numbers or {"base64": "..."}.'
        items: {}
        type: array
      command:
        description: A redis-cli style command line, alternative to args.
        type: string
      db:
        description: Database (standalone and sentinel).
        type: integer
      encoding:
        description: Encoding of binary strings in the reply.
        enum:
        - auto
        - utf8
        - base64
        type: string
      node:
        description: Node address for target=node.
        type: string
      target:
        description: 'Where the command goes: auto routes by key.'
        enum:
        - auto
        - node
        - masters
        - replicas
        - all
        - sentinels
        type: string
    type: object
...
  httpapi.ErrorResponse:
    properties:
      error_code:
        example: BAD_REQUEST
        type: string
      error_description:
        example: key is required
        type: string
      request_id:
        example: 6f1c2a9e0b7d4c3f8a5e1d2b3c4f5a6b
        type: string
    type: object
...
  httpapi.PipelineRequest:
    properties:
      atomic:
        description: Wrap the commands into MULTI/EXEC.
        type: boolean
      commands:
        description: 'Commands: arrays of arguments or command lines.'
        items:
          items: {}
          type: array
        type: array
      db:
        description: Database (standalone and sentinel).
        type: integer
      encoding:
        description: Encoding of binary strings in the reply.
        enum:
        - auto
        - utf8
        - base64
        type: string
      node:
        description: Node address for target=node.
        type: string
      session:
        description: 'Run on a dedicated connection: SELECT, AUTH, WATCH and MULTI
          are allowed.'
        type: boolean
      target:
        description: 'Where the command goes: auto routes by key.'
        enum:
        - auto
        - node
        - masters
        - replicas
        - all
        - sentinels
        type: string
    type: object
...
  httpapi.Response:
    properties:
      error_code:
        description: Code is "OK" on success or an error code.
        example: OK
        type: string
      error_description:
        example: ""
        type: string
      request_id:
        example: 6f1c2a9e0b7d4c3f8a5e1d2b3c4f5a6b
        type: string
      result: {}
    type: object
...
  service.CommandResult:
    properties:
      node:
        type: string
      truncated:
        type: boolean
      value: {}
    type: object
...
  service.ImportError:
    properties:
      error:
        type: string
      key: {}
      line:
        type: integer
    type: object
  service.ImportResult:
    properties:
      errors:
        items:
          $ref: '#/definitions/service.ImportError'
        type: array
      failed:
        type: integer
      imported:
        type: integer
      skipped:
        description: Skipped are existing keys (without replace) and keys exported
          with an error.
        type: integer
    type: object
...
  service.NodeInfo:
    properties:
      addr:
        type: string
      role:
        type: string
    type: object
...
  service.PipelineEntry:
    properties:
      error:
        type: string
      value: {}
    type: object
  service.PipelineResult:
    properties:
      node:
        type: string
      results:
        items:
          $ref: '#/definitions/service.PipelineEntry'
        type: array
      truncated:
        type: boolean
    type: object
...
  service.TopologyInfo:
    properties:
      commands:
        type: integer
      default_db:
        type: integer
      nodes:
        items:
          $ref: '#/definitions/service.NodeInfo'
        type: array
      sentinel_master:
        type: string
      topology:
        type: string
    type: object
info:
  contact: {}
  description: |-
    HTTP API for Redis and Valkey: standalone, cluster and sentinel.

    Every JSON response is wrapped into Response; errors use ErrorResponse with a stable error_code.

    Common query parameters (where they make sense):
    db - database (standalone and sentinel);
    node - send the request to the node with this address;
    target - auto (by key), node, masters, replicas, all or sentinels; fan-out targets return a result per node;
    encoding - auto (binary strings become {"base64": "..."}), utf8 or base64;
    key_encoding=base64 - keys, fields and members in the query string are base64;
    timeout - request timeout for blocking commands, e.g. 30s;
    duration - duration of streams, exports and imports;
    pretty - indented JSON.

    In request bodies keys and values are strings, numbers or {"base64": "..."} objects.
  license:
    name: MIT
  title: redigate API
  version: "1.0"
paths:
...
  /command:
    post:
      consumes:
      - application/json
      - text/plain
      description: |-
        Runs any command. The body is either JSON (CommandRequest) or a redis-cli style command line such as `SET key "a value"`.
        Commands changing connection state (SELECT, AUTH, HELLO, MULTI, WATCH, CLIENT REPLY, ...) run on a dedicated connection.
        SUBSCRIBE, MONITOR and SYNC are served by streaming endpoints. With a fan-out target (masters, replicas, all, sentinels) the result is service.FanOutResult.
      parameters:
      - description: JSON or a command line
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpapi.CommandRequest'
      - description: Node address
        in: query
        name: node
        type: string
      - description: masters, replicas, all or sentinels
        in: query
        name: target
        type: string
      - description: Database
        in: query
        name: db
        type: integer
      - description: auto, utf8 or base64
        in: query
        name: encoding
        type: string
      - description: Timeout for blocking commands, e.g. 30s
        in: query
        name: timeout
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.Response'
            - properties:
                result:
                  $ref: '#/definitions/service.CommandResult'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Run a command
      tags:
      - general
...
  /keys/export:
    get:
      description: 'Streams keys as NDJSON, one key per line (service.ExportRecord).
        format=dump keeps exact values with DUMP; format=value is portable between
        Redis and Valkey versions for core types. The last line is {"summary": {"keys":
        n, "complete": bool}}.'
      parameters:
      - description: Glob pattern
        in: query
        name: match
        type: string
      - description: Type filter
        in: query
        name: type
        type: string
      - description: dump or value
        in: query
        name: format
        type: string
      - description: Database
        in: query
        name: db
        type: integer
      - description: Maximum duration
        in: query
        name: duration
        type: string
      produces:
      - application/x-ndjson
      responses:
        "200":
          description: NDJSON, one service.ExportRecord per line
          schema:
            type: file
        default:
          description: ""
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export keys
      tags:
      - keys
  /keys/import:
    post:
      consumes:
      - application/x-ndjson
      description: Restores keys from NDJSON produced by export. The body may be up
        to MAX_IMPORT_BYTES.
      parameters:
      - description: NDJSON
        in: body
        name: request
        required: true
        schema:
          type: string
      - description: Overwrite existing keys
        in: query
        name: replace
        type: boolean
      - description: Database
        in: query
        name: db
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.Response'
            - properties:
                result:
                  $ref: '#/definitions/service.ImportResult'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Import keys
      tags:
      - keys
...
  /pipeline:
    post:
      consumes:
      - application/json
      - text/plain
      description: |-
        The body is JSON (PipelineRequest) or one redis-cli style command per line, lines starting with # are skipped.
        atomic wraps the commands into MULTI/EXEC; session runs them on a dedicated connection, which allows SELECT, AUTH, WATCH and manual MULTI/EXEC.
        In cluster atomic pipelines and sessions require keys of a single slot.
      parameters:
      - description: JSON or command lines
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpapi.PipelineRequest'
      - description: Node address
        in: query
        name: node
        type: string
      - description: Database
        in: query
        name: db
        type: integer
      - description: MULTI/EXEC
        in: query
        name: atomic
        type: boolean
      - description: Dedicated connection
        in: query
        name: session
        type: boolean
      - description: auto, utf8 or base64
        in: query
        name: encoding
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.Response'
            - properties:
                result:
                  $ref: '#/definitions/service.PipelineResult'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Run commands in one round trip
      tags:
      - general
...
  /topology:
    get:
      description: The detected topology (standalone, cluster or sentinel) and known
        nodes.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.Response'
            - properties:
                result:
                  $ref: '#/definitions/service.TopologyInfo'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Topology and nodes
      tags:
      - general
...
securityDefinitions:
  BearerAuth:
    description: '"Bearer <API_TOKEN>" or "Bearer <API_READONLY_TOKEN>". Not required
      when authentication is disabled.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"